If an exception occurs during SendErrorNotificationAsync, catch it and flag the error failure.
Also: if an HttpRequestException occurs during SendNewSetNotificationAsync, then consider the notification failed and DON'T attempt to send another notification regarding that error.

In other cases (if SendNewSetNotificationAsync doesn't fail, or if SendErrorNotificationAsync at least succeeds) consider the new set "seen" and *partially* update PreviouslySeenData.

---

Found today, due to a large volume of new sets being posted for January 2026: when multiple new sets are found, and an exception is throwing during one SendNewSetNotificationAsync, AND an exception is then thrown during SendErrorNotificationAsync -- the program itself throws and doesn't get to UpdateSetsAsync.

This results in the next LegoSetNotifier attempt seeing all the same new sets, and trying all of those notifications again.

The exception-then-exception-again case is especially likely when there's a large volume of new sets, because! rate-limiting from the notification target will result in the error notification then failing too.

Punchline is that if there are enough new sets at a time, that LegoSetNotifier run will hit a notification rate limit, fail to consider any of those sets "seen," then hit the exact same problem again on the next attempt.